StorageProject/requirements.txt
flask==3.0.3
sqlalchemy==2.0.30
mysql-connector-python==8.0.33
connexion==3.1.0
pyyaml==6.0.1
pymysql==1.1.1
connexion[flask]==3.1.0
connexion[uvicorn]==3.1.0
connexion[swagger-ui]==3.1.0
pykafka==2.8.0
Werkzeug==3.0.0
cryptography
